Battery animations for charging, ramp and low
So when charging it plays like it's filling, with a green underline
While discharging display a ramp (which seems to be just the capacity idly?) with the default orange-ish colour
When low (set it to 20), a red underline with a pretty quick success flashing between full and empty battery icons
Update full prefix to just full battery icon
Updated all icons to Font Awesome 6 (with more frames) as per https://github.com/polybar/polybar/wiki/Module:-battery
Looks pretty neat.
Updated animation-discharging even though not using that, because why would you want it to show an always draining down animation, at all?

Change icon to video icon for ws5 ( zoom)
And to tv on ws8 (stremio) in accordance with previous commit workspace assignments